+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to the "vscode-symfony-ux-twig-component" extension will be 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.2] - 2024-03-14
+
+- The formatter now works alongside other formatters
+- Completely redesigned configuration approach with base paths and relative component paths
+- Added support for multi-level namespaces in component tags
+- Improved navigation: now you can click on either the namespace or component name
+- Fixed attribute formatting issues
+
+## [0.0.1] - 2024-03-14
 
 - Initial release
+- Added syntax highlighting for Twig components
+- Added clickable namespaces for navigation
+- Added basic formatting for Twig component attributes